Use Pimple idiom for Network Interface Listener
Create `qnx` and `linux` folder with platform specific code.

Moved implementation from network_interface_listener_impl to
platform_specific/linux/platform_specific_network_interface_listener_impl

QNX contains empty implementation

Renamed source files : used cc file extension for source files
