+++ src/components/transport_manager/include/transport_manager/tcp/platform_specific/qnx/platform_specific_network_interface_listener_impl.h
@@ -30,14 +30,20 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_NETWORK_INTERFACE_LISTENER_IMPL_H_
-#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_NETWORK_INTERFACE_LISTENER_IMPL_H_
+#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_PLATFORM_SPECIFIC_QNX_PLATFORM_SPECIFIC_NETWORK_INTERFACE_LISTENER_H_
+#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_PLATFORM_SPECIFIC_QNX_PLATFORM_SPECIFIC_NETWORK_INTERFACE_LISTENER_H_
 
+#include <map>
 #include <string>
-#include <memory>
+#include <vector>
 
-#include "utils/macro.h"
+#include <netinet/in.h>
 #include "transport_manager/tcp/network_interface_listener.h"
+#include "utils/macro.h"
+#include "utils/threads/thread_delegate.h"
+
+class Thread;
+struct ifaddrmsg;
 
 namespace transport_manager {
 namespace transport_adapter {
@@ -47,7 +53,8 @@ class TcpClientListener;
 /**
  * @brief Listener to detect various events on network interfaces
  */
-class NetworkInterfaceListenerImpl : public NetworkInterfaceListener {
+class PlatformSpecificNetworkInterfaceListener
+    : public NetworkInterfaceListener {
  public:
   /**
    * @brief Constructor
@@ -57,13 +64,14 @@ class NetworkInterfaceListenerImpl : public NetworkInterfaceListener {
    * @param designated_interface  if we want to listen only on a specific
    *                              network interface, specify its name
    */
-  NetworkInterfaceListenerImpl(TcpClientListener* tcp_client_listener,
-                               const std::string designated_interface);
+  PlatformSpecificNetworkInterfaceListener(
+      TcpClientListener* tcp_client_listener,
+      const std::string designated_interface = "");
 
   /**
    * @brief Destructor
    */
-  virtual ~NetworkInterfaceListenerImpl();
+  virtual ~PlatformSpecificNetworkInterfaceListener();
 
   /**
    * @brief Initialize this listener
@@ -86,10 +94,9 @@ class NetworkInterfaceListenerImpl : public NetworkInterfaceListener {
   bool Stop() OVERRIDE;
 
  private:
-  std::unique_ptr<NetworkInterfaceListener> platform_specific_impl_;
 };
 
 }  // namespace transport_adapter
 }  // namespace transport_manager
 
-#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_NETWORK_INTERFACE_LISTENER_IMPL_H_
+#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_PLATFORM_SPECIFIC_QNX_PLATFORM_SPECIFIC_NETWORK_INTERFACE_LISTENER_H_
