+++ src/components/transport_manager/src/tcp/platform_specific/qnx/platform_specific_network_interface_listener.cc
@@ -29,67 +29,41 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-
-#ifndef SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_NETWORK_INTERFACE_LISTENER_IMPL_H_
-#define SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_NETWORK_INTERFACE_LISTENER_IMPL_H_
-
-#include <string>
-#include <memory>
-
-#include "utils/macro.h"
-#include "transport_manager/tcp/network_interface_listener.h"
+#include "transport_manager/tcp/platform_specific/qnx/platform_specific_network_interface_listener_impl.h"
 
 namespace transport_manager {
 namespace transport_adapter {
 
-class TcpClientListener;
+CREATE_LOGGERPTR_GLOBAL(logger_, "TransportManager")
 
-/**
- * @brief Listener to detect various events on network interfaces
- */
-class NetworkInterfaceListenerImpl : public NetworkInterfaceListener {
- public:
-  /**
-   * @brief Constructor
-   *
-   * @param tcp_client_listener  an instance of TcpClientListener which receives
-   *                             status updates
-   * @param designated_interface  if we want to listen only on a specific
-   *                              network interface, specify its name
-   */
-  NetworkInterfaceListenerImpl(TcpClientListener* tcp_client_listener,
-                               const std::string designated_interface);
+PlatformSpecificNetworkInterfaceListener::
+    PlatformSpecificNetworkInterfaceListener(
+        TcpClientListener* tcp_client_listener,
+        const std::string designated_interface) {}
 
-  /**
-   * @brief Destructor
-   */
-  virtual ~NetworkInterfaceListenerImpl();
+PlatformSpecificNetworkInterfaceListener::
+    ~PlatformSpecificNetworkInterfaceListener() {
+  LOG4CXX_AUTO_TRACE(logger_);
+}
 
-  /**
-   * @brief Initialize this listener
-   */
-  bool Init() OVERRIDE;
+bool PlatformSpecificNetworkInterfaceListener::Init() {
+  LOG4CXX_AUTO_TRACE(logger_);
+  return true;
+}
 
-  /**
-   * @brief Deinitialize this listener
-   */
-  void Deinit() OVERRIDE;
+void PlatformSpecificNetworkInterfaceListener::Deinit() {
+  LOG4CXX_AUTO_TRACE(logger_);
+}
 
-  /**
-   * @brief Start this listener
-   */
-  bool Start() OVERRIDE;
+bool PlatformSpecificNetworkInterfaceListener::Start() {
+  LOG4CXX_AUTO_TRACE(logger_);
+  return true;
+}
 
-  /**
-   * @brief Stop this listener
-   */
-  bool Stop() OVERRIDE;
-
- private:
-  std::unique_ptr<NetworkInterfaceListener> platform_specific_impl_;
-};
+bool PlatformSpecificNetworkInterfaceListener::Stop() {
+  LOG4CXX_AUTO_TRACE(logger_);
+  return true;
+}
 
 }  // namespace transport_adapter
 }  // namespace transport_manager
-
-#endif  // SRC_COMPONENTS_TRANSPORT_MANAGER_INCLUDE_TRANSPORT_MANAGER_TCP_NETWORK_INTERFACE_LISTENER_IMPL_H_
